requirements.txt
alembic==1.12.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==3.0.0
Flask-Login==0.6.2
flask-menu==1.0.0
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
WTForms==3.1.0
